Speed up filtering by orders of magnitude on big repositories
Previously i.e. on a Linux kernel repository it took many minutes using 100% of CPU, now it takes below a second. The problem was generation
of thousands of events about changed selection of gtk::TreeView (commit log), which happened during refiltering of the model, which triggered
more other events. Now signals are blocked during refiltering and selection is checked after unblocking if an event about it needs to be sent.