Redesign managing configuration, use toml+serde instead of rust-ini
Configuration no longer uses a system of "saver" objects that have to register into "settings". Instead, it uses event system to send changes to config store
and it is saved to a file just before quitting the application. Moreover, thanks to serde, it is more type-safe now.

Additionally, creation of repository has been moved to Gui struct because of better object compatibility with other "pub(crate)" types (Repository no longer is "pub").